+++ packages/xchain-aggregator/src/protocols/index.ts
@@ -1,2 +1,3 @@
 export { MayachainProtocol } from './mayachain'
 export { ThorchainProtocol } from './thorchain'
+export { ChainflipProtocol } from './chainflip'