+++ packages/xchain-mayachain/src/const.ts
@@ -26,6 +26,11 @@ export const MAYA_DENOM = 'maya'
 export const MAYA_DECIMAL = 4
 
 /**
+ * Number of decimals for the synthetic assets.
+ */
+export const SYNTH_DECIMAL = 8
+
+/**
  * Default gas limit value for transactions.
  */
 export const DEFAULT_GAS_LIMIT_VALUE = '4000000'
@@ -43,7 +48,7 @@ export const MAYAChain = 'MAYA' as const
 /**
  * Default fee for MayaChain transactions.
  */
-export const DEFAULT_FEE = baseAmount(5000000000, CACAO_DECIMAL)
+export const DEFAULT_FEE = baseAmount(2000000000, CACAO_DECIMAL)
 
 /**
  * Base "chain" asset on MayaChain main net.