+++ packages/xchain-wallet/src/types.ts
@@ -1,6 +1,6 @@
-import { TxParams as BaseTxParams } from '@xchainjs/xchain-client'
+import { Balance, TxParams as BaseTxParams } from '@xchainjs/xchain-client'
 import { TxParams as BaseEvmTxParams } from '@xchainjs/xchain-evm'
-import { Asset, BaseAmount } from '@xchainjs/xchain-util'
+import { Asset, BaseAmount, Chain } from '@xchainjs/xchain-util'
 
 /**
  * UTXO transfer params
@@ -10,3 +10,16 @@ export type UtxoTxParams = BaseTxParams & { asset: Asset; feeRate?: BaseAmount }
  * EVM transfer params
  */
 export type EvmTxParams = BaseEvmTxParams & { asset: Asset }
+
+/**
+ * Fulfilled balance from a Chain
+ */
+type FulfilledBalance = { status: 'fulfilled'; balances: Balance[] }
+/**
+ * Rejected balance from a Chain
+ */
+type RejectedBalance = { status: 'rejected'; reason: string }
+/**
+ * Balances by chain, which can be fulfilled or rejected
+ */
+export type ChainBalances = Record<Chain, FulfilledBalance | RejectedBalance>