+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.23.18
+
+### Patch Changes
+
+- Updated dependencies [8d000a2]
+  - @xchainjs/xchain-client@0.16.7
+  - @xchainjs/xchain-utxo@0.1.9
+  - @xchainjs/xchain-utxo-providers@0.2.17
+
 ## 0.23.17
 
 ### Patch Changes