+++ packages/xchain-cosmos-sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.2.12
+
+### Patch Changes
+
+- f29a7ad: New protected method `getChainId` to retrieve the network id the client is connected to
+- f29a7ad: New protected method `getAccount` to retrieve the account of a given address
+
+## 0.2.11
+
+### Patch Changes
+
+- 1a8f57d: Get transaction data bug fix
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [8d000a2]
+  - @xchainjs/xchain-client@0.16.7
+
 ## 0.2.9
 
 ### Patch Changes