+++ packages/xchain-mayachain-amm/CHANGELOG.md
@@ -1,5 +1,60 @@
 # Changelog
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [662085f]
+- Updated dependencies [662085f]
+- Updated dependencies [9d4c162]
+  - @xchainjs/xchain-mayachain-query@0.1.16
+  - @xchainjs/xchain-mayachain@1.0.10
+  - @xchainjs/xchain-wallet@0.1.18
+
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies [3863979]
+- Updated dependencies [f29a7ad]
+- Updated dependencies [3c6a033]
+- Updated dependencies [a9fd322]
+  - @xchainjs/xchain-wallet@0.1.17
+  - @xchainjs/xchain-thorchain@1.1.0
+  - @xchainjs/xchain-kujira@0.1.20
+  - @xchainjs/xchain-mayachain@1.0.9
+
+## 2.0.7
+
+### Patch Changes
+
+- @xchainjs/xchain-kujira@0.1.19
+- @xchainjs/xchain-mayachain@1.0.8
+- @xchainjs/xchain-thorchain@1.0.11
+- @xchainjs/xchain-wallet@0.1.16
+
+## 2.0.6
+
+### Patch Changes
+
+- @xchainjs/xchain-ethereum@0.32.5
+- @xchainjs/xchain-wallet@0.1.15
+
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies [8d000a2]
+  - @xchainjs/xchain-client@0.16.7
+  - @xchainjs/xchain-bitcoin@0.23.18
+  - @xchainjs/xchain-dash@0.3.5
+  - @xchainjs/xchain-ethereum@0.32.4
+  - @xchainjs/xchain-kujira@0.1.18
+  - @xchainjs/xchain-mayachain@1.0.7
+  - @xchainjs/xchain-mayachain-query@0.1.15
+  - @xchainjs/xchain-thorchain@1.0.10
+  - @xchainjs/xchain-wallet@0.1.14
+
 ## 2.0.4
 
 ### Patch Changes