+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.7.14
+
+### Patch Changes
+
+- 662085f: `Swap` type updated, `outboundTx` property is now optional
+- 662085f: Outbound transaction bug fix in `getSwapHistory` method.
+
+## 0.7.13
+
+### Patch Changes
+
+- Updated dependencies [8d000a2]
+  - @xchainjs/xchain-client@0.16.7
+  - @xchainjs/xchain-midgard-query@0.1.18
+
 ## 0.7.12
 
 ### Patch Changes