+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## 1.1.0
+
+### Minor Changes
+
+- f29a7ad: Thorchain Ledger support
+
+### Patch Changes
+
+- Updated dependencies [f29a7ad]
+- Updated dependencies [f29a7ad]
+  - @xchainjs/xchain-cosmos-sdk@0.2.12
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [1a8f57d]
+  - @xchainjs/xchain-cosmos-sdk@0.2.11
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [8d000a2]
+  - @xchainjs/xchain-client@0.16.7
+  - @xchainjs/xchain-cosmos-sdk@0.2.10
+
 ## 1.0.9
 
 ### Patch Changes